Made syscall ids 64 bits in rax

+++ src/kernel/syscall.h
@@ -1,6 +1,6 @@
 #pragma once
 
-enum class syscall : uint16_t
+enum class syscall : uint64_t
 {
 	debug,
 	message,
